+++ docs/components/plugins/api/APISectionUtils.tsx
@@ -12,7 +12,6 @@ import {
   CommentData,
   MethodParamData,
   TypeDefinitionData,
-  TypeDefinitionTypesData,
 } from '~/components/plugins/api/APIDataTypes';
 
 export enum TypeDocKind {
@@ -53,14 +52,15 @@ export const mdInlineRenderers: MDRenderers = {
 
 const nonLinkableTypes = ['Date', 'T', 'TaskOptions', 'Uint8Array'];
 
-export const resolveTypeName = ({
-  elementType,
-  name,
-  type,
-  types,
-  typeArguments,
-  declaration,
-}: TypeDefinitionData): string | JSX.Element | (string | JSX.Element)[] => {
+export const resolveTypeName = (d: TypeDefinitionData): string | JSX.Element | (string | JSX.Element)[] => {
+  const {
+    elementType,
+    name,
+    type,
+    types,
+    typeArguments,
+    declaration,
+  } = d;
   if (name) {
     if (type === 'reference') {
       if (typeArguments) {
@@ -108,7 +108,7 @@ export const resolveTypeName = ({
     return elementType.name + type;
   } else if (type === 'union' && types?.length) {
     return types
-      .map((t: TypeDefinitionTypesData) =>
+      .map((t: TypeDefinitionData) =>
         t.type === 'reference' ? (
           <Link href={`#${t.name?.toLowerCase()}`} key={`type-link-${t.name}`}>
             {t.name}
@@ -136,14 +136,18 @@ export const resolveTypeName = ({
           {baseSignature.parameters?.map((param, index) => (
             <span key={`param-${index}-${param.name}`}>
               {param.name}: {resolveTypeName(param.type)}
-              {index + 1 !== baseSignature.parameters.length && ', '}
+              {index + 1 !== baseSignature.parameters?.length && ', '}
             </span>
           ))}
           ) {'=>'} {resolveTypeName(baseSignature.type)}
         </>
       );
     } else {
-      return `() => ${resolveTypeName(baseSignature.type)}`;
+      return (
+        <>
+          {'() =>'} {resolveTypeName(baseSignature.type)}
+        </>
+      );
     }
   }
   return 'undefined';